Register the factory menu items under Tracks
