+++ src/menus/LabelMenus.cpp
@@ -8,7 +8,8 @@
 #include "../ProjectAudioIO.h"
 #include "../ProjectHistory.h"
 #include "../ProjectSettings.h"
-#include "../SelectFile.h"
+#include "../ProjectWindow.h"
+#include "../SelectUtilities.h"
 #include "../SyncLock.h"
 #include "../TrackPanelAx.h"
 #include "../TrackPanel.h"
@@ -672,6 +673,25 @@ void OnDisjoinLabels(const CommandContext &context)
       XO( "Detach Labeled Audio" ) );
 }
 
+void OnNewLabelTrack(const CommandContext &context)
+{
+   auto &project = context.project;
+   auto &tracks = TrackList::Get( project );
+   auto &window = ProjectWindow::Get( project );
+
+   auto t = tracks.Add( std::make_shared<LabelTrack>() );
+
+   SelectUtilities::SelectNone( project );
+
+   t->SetSelected(true);
+
+   ProjectHistory::Get( project )
+      .PushState(XO("Created new label track"), XO("New Track"));
+
+   TrackFocus::Get(project).Set(t);
+   t->EnsureVisible();
+}
+
 }; // struct Handler
 
 } // namespace
@@ -795,6 +815,12 @@ AttachedItem sAttachment1{
    Shared( LabelEditMenus() )
 };
 
-}
+AttachedItem sAttachment2{ wxT("Tracks/Add/Add"),
+   ( FinderScope{ findCommandHandler },
+   Command( wxT("NewLabelTrack"), XXO("&Label Track"),
+      FN(OnNewLabelTrack), AudioIONotBusyFlag() )
+   )
+};
 
+}
 #undef FN
