+++ src/menus/TrackMenus.cpp
@@ -15,7 +15,6 @@
 #include "../SelectUtilities.h"
 #include "../ShuttleGui.h"
 #include "../SyncLock.h"
-#include "../TimeTrack.h"
 #include "../TrackPanelAx.h"
 #include "../TrackPanel.h"
 #include "../TrackUtilities.h"
@@ -609,103 +608,6 @@ namespace TrackActions {
 // Menu handler functions
 
 struct Handler : CommandHandlerObject {
-
-void OnNewWaveTrack(const CommandContext &context)
-{
-   auto &project = context.project;
-   auto &tracks = TrackList::Get( project );
-   auto &trackFactory = WaveTrackFactory::Get( project );
-   auto &window = ProjectWindow::Get( project );
-
-   auto defaultFormat = QualitySettings::SampleFormatChoice();
-
-   auto rate = ProjectRate::Get(project).GetRate();
-
-   auto t = tracks.Add( trackFactory.NewWaveTrack( defaultFormat, rate ) );
-   SelectUtilities::SelectNone( project );
-
-   t->SetSelected(true);
-
-   ProjectHistory::Get( project )
-      .PushState(XO("Created new audio track"), XO("New Track"));
-
-   TrackFocus::Get(project).Set(t);
-   t->EnsureVisible();
-}
-
-void OnNewStereoTrack(const CommandContext &context)
-{
-   auto &project = context.project;
-   auto &tracks = TrackList::Get( project );
-   auto &trackFactory = WaveTrackFactory::Get( project );
-   auto &window = ProjectWindow::Get( project );
-
-   auto defaultFormat = QualitySettings::SampleFormatChoice();
-   auto rate = ProjectRate::Get(project).GetRate();
-
-   SelectUtilities::SelectNone( project );
-
-   auto left = tracks.Add( trackFactory.NewWaveTrack( defaultFormat, rate ) );
-   left->SetSelected(true);
-
-   auto right = tracks.Add( trackFactory.NewWaveTrack( defaultFormat, rate ) );
-   right->SetSelected(true);
-
-   tracks.MakeMultiChannelTrack(*left, 2, true);
-
-   ProjectHistory::Get( project )
-      .PushState(XO("Created new stereo audio track"), XO("New Track"));
-
-   TrackFocus::Get(project).Set(left);
-   left->EnsureVisible();
-}
-
-void OnNewLabelTrack(const CommandContext &context)
-{
-   auto &project = context.project;
-   auto &tracks = TrackList::Get( project );
-   auto &trackFactory = WaveTrackFactory::Get( project );
-   auto &window = ProjectWindow::Get( project );
-
-   auto t = tracks.Add( std::make_shared<LabelTrack>() );
-
-   SelectUtilities::SelectNone( project );
-
-   t->SetSelected(true);
-
-   ProjectHistory::Get( project )
-      .PushState(XO("Created new label track"), XO("New Track"));
-
-   TrackFocus::Get(project).Set(t);
-   t->EnsureVisible();
-}
-
-void OnNewTimeTrack(const CommandContext &context)
-{
-   auto &project = context.project;
-   auto &tracks = TrackList::Get( project );
-   auto &viewInfo = ViewInfo::Get( project );
-   auto &window = ProjectWindow::Get( project );
-
-   if ( *tracks.Any<TimeTrack>().begin() ) {
-      AudacityMessageBox(
-         XO(
-"This version of Audacity only allows one time track for each project window.") );
-      return;
-   }
-
-   auto t = tracks.AddToHead( std::make_shared<TimeTrack>(&viewInfo) );
-
-   SelectUtilities::SelectNone( project );
-
-   t->SetSelected(true);
-
-   ProjectHistory::Get( project )
-      .PushState(XO("Created new time track"), XO("New Track"));
-
-   TrackFocus::Get(project).Set(t);
-   t->EnsureVisible();
-}
 
 void OnStereoToMono(const CommandContext &context)
 {
@@ -1308,16 +1210,7 @@ BaseItemSharedPtr TracksMenu()
    ( FinderScope{ findCommandHandler },
    Menu( wxT("Tracks"), XXO("&Tracks"),
       Section( "Add",
-         Menu( wxT("Add"), XXO("Add &New"),
-            Command( wxT("NewMonoTrack"), XXO("&Mono Track"), FN(OnNewWaveTrack),
-               AudioIONotBusyFlag(), wxT("Ctrl+Shift+N") ),
-            Command( wxT("NewStereoTrack"), XXO("&Stereo Track"),
-               FN(OnNewStereoTrack), AudioIONotBusyFlag() ),
-            Command( wxT("NewLabelTrack"), XXO("&Label Track"),
-               FN(OnNewLabelTrack), AudioIONotBusyFlag() ),
-            Command( wxT("NewTimeTrack"), XXO("&Time Track"),
-               FN(OnNewTimeTrack), AudioIONotBusyFlag() )
-         )
+         Menu( wxT("Add"), XXO("Add &New") )
       ),
 
       //////////////////////////////////////////////////////////////////////////
