+++ src/tracks/playabletrack/notetrack/ui/NoteTrackMenuItems.cpp
@@ -1,0 +1,15 @@
+/**********************************************************************
+
+Audacity: A Digital Audio Editor
+
+@file NoteTrackMenuItems.cpp
+@brief Injects menu items using NoteTrack but not the views of it
+
+Paul Licameli split from TrackMenus.cpp
+
+**********************************************************************/
+
+/*!
+ There is no "New Note Track" menu item corresponding to the other track types.  Note tracks are created
+ when importing MIDI files.
+ */
